Refactor S3FileSystem Registration
